Move spinner.py into views/components #48
